server/ai-microservice/requirements.txt
flask
pymongo[srv]
sentence-transformers
faiss-cpu
numpy
python-dotenv
